PURR OS v0.6.0 — MajorasMask
- explorer.paws: Windows CE-inspired app launcher shell with dynamic icon
  grid (populated from kitt.app_list), silver taskbar, left strip, PURR OS
  watermark, green LED heartbeat in normal mode
- purr_bootloader: redesigned OTA manager — OTA 0 = PURR OS (reinstall +
  remove with warning), OTA 1 = custom firmware (launch/install/delete);
  bootloader_only NVS flag locks device into bootloader until PURR OS is
  reinstalled; flickering fixed via surgical header RAM refresh
- touch_cst816s: new CST816S capacitive I2C touch driver for ESP32-2432S024C
  (SDA=33, SCL=32, INT=21, RST=25); replaces XPT2046 for S024C variant
- display_ili9341: backlight GPIO variant support (S024C=27, S028R=21);
  LEDC-only init removes flicker on startup; drawing primitives added
- power_manager: BATT_ADC_PIN=-1 guard for ILI9341 targets (GPIO 1 crash fix)
- Boot routing: GPIO 0 hold OR NVS boot_mode/bootloader_only flag enters
  bootloader; explorer is normal boot target; .meow/.paws app type convention
- CMakeLists: purr_shell renamed to purr_bootloader + explorer, CST816S driver
  wired in, PURR_HAS_BOOTLOADER + PURR_HAS_EXPLORER flags, CYD_VARIANT_S024C
- Build.ps1: SPIFFS image builder, Espressif PS profile detection fixed,
  esptool flash command with all five regions including SPIFFS at 0x3b0000

+++ CoreOS/system/kernel/modules/partition_manager.cpp
@@ -183,7 +183,7 @@ bool pm_install(uint8_t slot, const char* sd_path, const char* display_name,
 
     size_t filesize = f.size();
     if (filesize == 0 || filesize > part->size) {
-        Serial.printf("[pm] install: bad size %u (part %u)\n", filesize, part->size);
+        Serial.printf("[pm] install: bad size %lu (part %lu)\n", (unsigned long)filesize, (unsigned long)part->size);
         f.close();
         return false;
     }
